Removed Streams api and moved methods directly to EventStoreDBClient
