+++ db-client-java/src/test/java/com/eventstore/dbclient/SubscribeToStreamTests.java
@@ -16,7 +16,7 @@ public class SubscribeToStreamTests {
 
     @Test
     public void testStreamSubscriptionDeliversAllowsCancellationDuringStream() throws InterruptedException, ExecutionException {
-        Streams streams = server.getStreamsAPI();
+        EventStoreDBClient client = server.getClient();
 
         final CountDownLatch receivedEvents = new CountDownLatch(1000);
         final CountDownLatch cancellation = new CountDownLatch(1);
@@ -38,7 +38,7 @@ public class SubscribeToStreamTests {
             }
         };
 
-        Subscription result = streams.subscribeToStream("dataset20M-0", listener)
+        Subscription result = client.subscribeToStream("dataset20M-0", listener)
                 .get();
 
         receivedEvents.await();
@@ -48,7 +48,7 @@ public class SubscribeToStreamTests {
 
     @Test
     public void testStreamSubscriptionDeliversAllEventsInStream() throws InterruptedException, ExecutionException {
-        Streams streams = server.getStreamsAPI();
+        EventStoreDBClient client = server.getClient();
 
         final CountDownLatch receivedEvents = new CountDownLatch(6000);
         final CountDownLatch cancellation = new CountDownLatch(1);
@@ -76,7 +76,7 @@ public class SubscribeToStreamTests {
 
         CountingListener listener = new CountingListener();
 
-        Subscription result = streams.subscribeToStream("dataset20M-0", listener)
+        Subscription result = client.subscribeToStream("dataset20M-0", listener)
                 .get();
 
         receivedEvents.await();
@@ -86,7 +86,7 @@ public class SubscribeToStreamTests {
 
     @Test
     public void testStreamSubscriptionDeliversAllEventsInStreamAndListensForNewEvents() throws Throwable {
-        Streams streams = server.getStreamsAPI();
+        EventStoreDBClient client = server.getClient();
 
         final CountDownLatch receivedEvents = new CountDownLatch(6000);
         final CountDownLatch appendedEvents = new CountDownLatch(1);
@@ -137,7 +137,7 @@ public class SubscribeToStreamTests {
         // Listen to everything already in the stream
         CountingListener listener = new CountingListener();
 
-        Subscription subscription = streams.subscribeToStream(testStreamName, listener)
+        Subscription subscription = client.subscribeToStream(testStreamName, listener)
                 .get();
 
         receivedEvents.await();
@@ -151,7 +151,7 @@ public class SubscribeToStreamTests {
         AppendToStreamOptions options = AppendToStreamOptions.get()
                 .expectedRevision(ExpectedRevision.expectedRevision(5999));
 
-        WriteResult writeResult = streams.appendToStream(testStreamName, options, event)
+        WriteResult writeResult = client.appendToStream(testStreamName, options, event)
                 .get();
 
         assertEquals(new StreamRevision(6000), writeResult.getNextExpectedRevision());
